trade_bot/requirements.txt
ccxt==4.1.96
pandas==2.0.3
python-dateutil==2.8.2
requests==2.31.0
scikit-learn==1.3.0
matplotlib==3.7.2
numpy==1.24.3
talib-binary==0.4.28